refactor: fake 클래스 패키지 이동, cron 값 inline
