+++ backend/src/main/resources/application.yml
@@ -33,9 +33,6 @@ logging:
     org.hibernate.type: trace
     org.hibernate.sql: debug
 
-schedules:
-  round-update: 0 0 0 * * *
-
 oauth2:
   github:
     client:
