+++ backend/src/test/resources/application.yml
@@ -25,9 +25,6 @@ logging:
     org.hibernate.type: trace
     org.hibernate.sql: debug
 
-schedules:
-  round-update: 0 0 0 * * *
-
 oauth2:
   github:
     client:
